Tweak travis ruby versions
When you don't specify the minior version of Ruby, travis will use the
latest minior and patch release of that version (ie with "2.3", travis
will run version 2.3.1 as it's the latest minor 2.3 Ruby).

I also added other versions to the matrix to test against for people who
might be stuck in old Ruby versions. If nothing else, they're a sanity
check and shows what versions of Ruby this gem should be able to pass
tests against.

NOTE: At this time, Travis seems to not understand that "2.3" can be
fuzzy searched with rvm to get 2.3.1, so the version is currently
hardcoded for now.